@@ -1,9 +1,9 @@
 name: random_spin
 description: Random Spin App.
 
-publish_to: 'none' # Remove this line if you wish to publish to pub.dev
+publish_to: 'none'
 
-version: 3.0.0+4
+version: 3.1.0+5
 
 environment:
   sdk: '>=2.19.6 <3.0.0'
@@ -14,26 +14,26 @@ dependencies:
   flutter_localizations:
     sdk: flutter
 
-  cupertino_icons: ^1.0.2
-  get: ^4.6.5
-  shared_preferences: ^2.1.0
-  flutter_fortune_wheel: ^1.2.0
-  audioplayers: ^4.0.1
-  package_info_plus: ^3.1.2
-  url_launcher: ^6.1.10
-  font_awesome_flutter: ^10.4.0
-  flutter_svg: ^2.0.5
-  google_mobile_ads: ^3.0.0
-  flutter_native_splash: ^2.2.19
-  intl: ^0.18.1
-  upgrader: ^6.5.0
-  loading_plus: ^1.3.0
+  cupertino_icons: ^1.0.8
+  get: ^4.6.6
+  shared_preferences: ^2.3.0
+  flutter_fortune_wheel: ^1.3.1
+  audioplayers: ^6.0.0
+  package_info_plus: ^8.0.1
+  url_launcher: ^6.3.0
+  font_awesome_flutter: ^10.7.0
+  flutter_svg: ^2.0.10+1
+  google_mobile_ads: ^5.1.0
+  flutter_native_splash: ^2.4.1
+  intl: ^0.19.0
+  upgrader: ^10.3.0
+  loading_plus: ^2.0.2
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  flutter_lints: ^2.0.0
+  flutter_lints: ^4.0.0
   flutter_launcher_icons: ^0.13.1
   import_sorter: ^4.6.0
 
@@ -41,7 +41,7 @@ import_sorter:
   emojis: false
   comments: false
 
-#pub run import_sorter:main
+# dart run import_sorter:main
 
 flutter_launcher_icons:
   android: "launcher_icon"
