@@ -1,23 +1,4 @@
-import 'package:flutter/material.dart';
-
-import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-import 'package:get/get.dart';
-import 'package:google_mobile_ads/google_mobile_ads.dart';
-import 'package:loading_plus/loading_plus.dart';
-import 'package:package_info_plus/package_info_plus.dart';
-import 'package:shared_preferences/shared_preferences.dart';
-
-import 'package:random_spin/modules/saved_lists/view/saved_lists_view.dart';
-import 'package:random_spin/modules/settings/bindings/settings_binding.dart';
-import 'package:random_spin/repositories/language/language_repository.dart';
-import 'modules/about_us/view/about_us_view.dart';
-import 'modules/home/binding/home_binding.dart';
-import 'modules/home/view/home_view.dart';
-import 'modules/saved_lists/bindings/saved_lists_binding.dart';
-import 'modules/settings/view/settings_view.dart';
-import 'modules/use_instructions/view/use_instructions_view.dart';
-import 'repositories/language/language_locale.dart';
-import 'utils/resources/theme_manager.dart';
+import 'utils/all_imports.dart';
 
 late SharedPreferences sharedPreferences;
 late PackageInfo packageInfo;
@@ -33,7 +14,6 @@ Future<void> main() async {
   );
   sharedPreferences = await SharedPreferences.getInstance();
   packageInfo = await PackageInfo.fromPlatform();
-  LoadingPlus.instance.init(navigatorKey);
   runApp(MyApp());
 }
 
@@ -55,6 +35,11 @@ class MyApp extends StatelessWidget {
       locale: Locale(languageRepository.getCurrentLanguage()),
       navigatorKey: navigatorKey,
       theme: getApplicationTheme(),
+      builder: (BuildContext context, Widget? child) {
+        return LoadingPlus(
+          child: child!,
+        );
+      },
       initialRoute: HomeView.routeName,
       initialBinding: SettingsBinding(),
       getPages: [
