@@ -1,7 +1,4 @@
-import 'package:get/get.dart';
-
-import 'package:random_spin/modules/home/controller/home_controller.dart';
-import 'package:random_spin/repositories/saved_lists/saved_lists_locale.dart';
+import '../../../utils/all_imports.dart';
 
 class HomeBinding implements Bindings {
   @override
