@@ -1,7 +1,4 @@
-import 'package:get/get.dart';
-
-import 'package:random_spin/repositories/saved_lists/saved_lists_locale.dart';
-import '../controller/saved_lists_controller.dart';
+import '../../../utils/all_imports.dart';
 
 class SavedListsBinding extends Bindings {
   @override
