@@ -1,6 +1,4 @@
-import 'package:random_spin/utils/resources/storage_manager.dart';
-import '../../main.dart';
-import 'sound_repository.dart';
+import '../../../utils/all_imports.dart';
 
 class SoundLocale extends SoundRepository {
   @override
